Adding books and readers

+++ src/main/java/pl/wietwioorki/to22019/model/Author.java
@@ -6,7 +6,7 @@ import lombok.*;
 @Getter
 @ToString
 public class Author {
-    private int authorId;
+    private Long authorId;
     private String fullName;
 //    private List<Book> writtenBooks;
 }
